+++ Asterix/main_THD.py
@@ -4,7 +4,7 @@
 import os
 
 from Asterix.utils import read_parameter_file
-from Asterix.optical_systems import Pupil, Coronagraph, DeformableMirror, Testbed
+from Asterix.optics import Pupil, Coronagraph, DeformableMirror, Testbed
 from Asterix.wfsc import Estimator, Corrector, MaskDH
 from Asterix.loop import correction_loop, save_loop_results
