--- Asterix/optical_systems/__init__.py
@@ -1,13 +1,0 @@
-__all__ = [
-    'OpticalSystem',
-    'Pupil',
-    'Coronagraph',
-    'DeformableMirror',
-    'Testbed'
-]
-
-from .optical_systems import *
-from .pupil import *
-from .coronagraph import *
-from .deformable_mirror import *
-from .testbed import *
