+++ Asterix/optics/__init__.py
@@ -1,2 +1,16 @@
+__all__ = [
+    'OpticalSystem',
+    'Pupil',
+    'Coronagraph',
+    'DeformableMirror',
+    'Testbed'
+]
+
+from .optical_systems import *
+from .pupil import *
+from .coronagraph import *
+from .deformable_mirror import *
+from .testbed import *
+
 from .phase_amplitude_functions import *
 from .propagation_functions import *
