+++ Asterix/optics/optical_systems.py
@@ -6,7 +6,6 @@ import numpy as np
 from astropy.io import fits
 
 from Asterix import model_dir
-
 import Asterix.optics.propagation_functions as prop
 import Asterix.optics.phase_amplitude_functions as phase_ampl
 from Asterix.utils import save_plane_in_fits, ft_subpixel_shift, ft_zoom_out, crop_or_pad_image
