+++ Asterix/optics/testbed.py
@@ -5,7 +5,7 @@ import inspect
 import copy
 import numpy as np
 
-from Asterix.optical_systems import OpticalSystem, DeformableMirror
+from Asterix.optics import OpticalSystem, DeformableMirror
 
 
 class Testbed(OpticalSystem):
