+++ Asterix/wfsc/estimator.py
@@ -6,7 +6,7 @@ import numpy as np
 from astropy.io import fits
 
 from Asterix.utils import resizing
-from Asterix.optical_systems import OpticalSystem, DeformableMirror, Testbed
+from Asterix.optics import OpticalSystem, DeformableMirror, Testbed
 import Asterix.wfsc.wf_sensing_functions as wfs
