+++ Asterix/wfsc/wf_control_functions.py
@@ -8,9 +8,10 @@ matplotlib.use('TkAgg')
 import matplotlib.pyplot as plt
 from astropy.io import fits
 
-import Asterix.optics.propagation_functions as prop
+
 from Asterix.utils import resizing, crop_or_pad_image, save_plane_in_fits, progress
-from Asterix.optical_systems import OpticalSystem, DeformableMirror, Testbed
+import Asterix.optics.propagation_functions as prop
+from Asterix.optics import OpticalSystem, DeformableMirror, Testbed
 
 
 def create_interaction_matrix(testbed: Testbed,
